Implementing chat room lists
