+++ src/main/java/com/itbank/smartFarm/controller/HomeController.java
@@ -3,7 +3,6 @@ package com.itbank.smartFarm.controller;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
 
 @Controller
 public class HomeController {
@@ -11,7 +10,6 @@ public class HomeController {
     @GetMapping("/")
     public String home(Model model) {
     	model.addAttribute("msg", "Project Start");
-    	
         return "home";
     }
 
@@ -24,10 +22,5 @@ public class HomeController {
     @GetMapping("/support")
     public void support(Model model) {}
 
-    @GetMapping("/chat/chatScreen/{senderId}/{receiverId}")
-    public String chat(Model model, @PathVariable String senderId, @PathVariable String receiverId) {
-        model.addAttribute("sender_id", senderId);
-        model.addAttribute("receiver_id", receiverId);
-        return "/chat/chatScreen";
-    }
+
 }
