+++ src/main/resources/templates/chat/chatList.html
@@ -1,0 +1,17 @@
+<!DOCTYPE html>
+<html lang="en" xmlns:th="http://www.thymeleaf.org">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Chat List</title>
+</head>
+<body>
+    <div id="chatList">
+        <ul>
+            <li th:each="sender : ${senders}">
+                <a th:href="@{/chat/chatScreen(receiverId=${sender.id})}" th:text="${sender.nick}"></a>
+            </li>
+        </ul>
+    </div>
+</body>
+</html>
